Add stub methods in DataServiceMostRecentSeason
